refactor: clean up file hierarchy
