+++ src/modules/datastore/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./datastore";
+export * from "./local-storage-datastore";
+export * as Utilities from "./utilities";
