+++ src/modules/datastore/local-storage-datastore.ts
@@ -1,32 +1,4 @@
-export interface DataStore<T> {
-    entries: (Entry & T)[];
-
-    get(entry: Entry): Promise<Entry & T>;
-    add(entry: T): Promise<Entry & T>;
-    update(entry: Entry & Partial<T>): Promise<void>;
-    remove(entry: Entry): Promise<void>;
-
-    onChange(update: (store: DataStore<T>) => void);
-    offChange(update: (store: DataStore<T>) => void);
-}
-
-export type Entry = {
-    id: string;
-}
-
-export type Post = {
-    title: string;
-    type: string;
-    content: string;
-    author?: string;
-    imageUrl?: string;
-    index?: number,
-}
-
-export type Author = {
-    name: string;
-    index?: number;
-}
+import {Entry, DataStore} from "./datastore";
 
 export class LocalStorageStore<T> implements DataStore<T> {
     private _targetId: string;
