+++ src/modules/datastore/types.ts
@@ -1,0 +1,13 @@
+export type Post = {
+    title: string;
+    type: string;
+    content: string;
+    author?: string;
+    imageUrl?: string;
+    index?: number,
+}
+
+export type Author = {
+    name: string;
+    index?: number;
+}
